Not Found Exception Added

+++ 08.Clean architecture/BlogApi/src/Core/BlogApi.Application/Features/Post/Handlers/Commands/DeletePostCommandHandler.cs
@@ -3,6 +3,8 @@ using AutoMapper;
 using MediatR;
 using BlogApi.Application.Features.Post.Requests;
 using BlogApi.Application.Persistence.Contracts;
+using BlogApi.Application.Exceptions;
+
 namespace BlogApi.Application.Features.Post.Handlers.Commands;
 
 public class DeletePostCommandHandler : IRequestHandler<DeletePostRequest, Unit>
@@ -18,6 +20,10 @@ public class DeletePostCommandHandler : IRequestHandler<DeletePostRequest, Unit>
     public async Task<Unit> Handle(DeletePostRequest request, CancellationToken cancellationToken)
     {
         var post = await _postRepository.Get(request.Id);
+        if (post == null)
+        {
+            throw new NotFoundException(nameof(Post), request.Id);
+        }
         await _postRepository.Delete(post);
         return Unit.Value;
     }
